Features 23 (#47)
* Benchmark test

Quick end-to-end test file to use Jconsole to help find a bottleneck in UserComputeAPI implementations.
Generates a large input payload, runs both implementations and prints the results in ms.

* Created PowerDigitSumControllerFast

Added a Cache version of PowerDigitSumController to speed up processing time.

* UnserComputeSpeedIT to changed to also test PowerDigitSumControllerFast

Updated UnserComputeSpeedIT to test to compare the speed of PowerDigitSumController with PowerDigitSumControllerFast.

Added a comment at the top of PowerDigitSumControllerFast explaining step taken to identify and fix bottleneck.

* Update README.md

 Document the results of the test Included a before and after benchmark numbers. Included a link to the benchmark tests.